+++ src/rendering/entities_rendering.rs
@@ -55,10 +55,10 @@ fn draw_item(d: &mut RaylibDrawHandle, item: &dyn Entity, engine: &GameEngine) {
         };
 
         let actual_col = frame.x as f32 - engine.camera_viewport.x as f32;
-        let actual_offset_x = offset.x as f32 - engine.camera_viewport_offset.x as f32;
+        let actual_offset_x = offset.x - engine.camera_viewport_offset.x;
 
         let actual_row = frame.y as f32 - engine.camera_viewport.y as f32;
-        let actual_offset_y = offset.y as f32 - engine.camera_viewport_offset.y as f32;
+        let actual_offset_y = offset.y - engine.camera_viewport_offset.y;
 
         let dest_rect = Rectangle {
             x: actual_col * tile_scale + actual_offset_x * scale,
